# docs/openapi.yaml
openapi: 3.0.3
info:
  title: АИС «Реестр обязательных требований» — HTTP API
  version: "1.0.0"
  description: |
    Спецификация фактического API веб-приложения (Next.js App Router, 46 маршрутов).
    Составлена по коду `src/app/api/**` (2026-07-02), см. docs/architecture (комплект ГОСТ 34).

    **Аутентификация:** JWT в httpOnly-cookie `auth_token` (выдаётся `/api/auth/login`, срок 7 суток).
    **Роли:** `admin` (супер-админ МНЭ) · `mne` (сотрудник МНЭ: всё как admin, кроме выгрузки каталога и создания учёток уровня МНЭ) · `moderator` (орган + поддерево организаций) ·
    `expert` (назначенные сферы ∩ органы). Скоупы возвращает `/api/auth/me`.
    **Ошибки:** единый формат `{ "error": string }`; 401 — не авторизован, 403 — нет прав/чужой орган,
    404 — не найдено, 429 — превышен лимит (голосование), 5xx — внутренняя/внешний сервис.

    Раздел «Легаси-экспертиза» обслуживает исторический MVP-контур (requirement_cards/голосования)
    и не развивается — см. документ 09 плана (К5).
servers:
  - url: https://reestr-web-production.up.railway.app
    description: Продуктив (Railway)
  - url: http://localhost:3000
    description: Локальная разработка
security:
  - cookieAuth: []
...
paths:
...
  /api/registry/npa:
    get:
      tags: [Реестр]
      summary: НПА узла органа (по authority_code требований; скоуп для не-admin)
      parameters:
        - { name: authority, in: query, required: true, schema: { type: string }, description: Код узла organizations }
        - { name: q, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { type: string }, description: Статус НПА (действующий/утратил силу) }
      responses:
        "200": { description: Список НПА со счётчиками и adilet-ссылками }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /api/npa-admin:
    get:
      tags: [Администрирование]
      summary: "Управление НПА (только admin): список НПА органа или статьи одного НПА"
      description: >
        Без ngr — НПА корневого органа (подразделение-путь, счётчики, действующие
        назначения). С ngr и articles=1 — разбивка НПА по статьям (метка, нормы,
        исключено, текущий орган).
      parameters:
        - { name: org_id, in: query, schema: { type: integer }, description: Корневой орган (без ngr — обязателен) }
        - { name: q, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: ngr, in: query, schema: { type: string } }
        - { name: articles, in: query, schema: { type: string, enum: ["1"] } }
      responses:
        "200": { description: "items[], log[] | articles[]" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [Администрирование]
      summary: "Передать НПА/статьи любому органу; исключить или вернуть (только admin)"
      description: >
        transfer — межминистерский каскад authority_code (целиком или по меткам статей;
        пересекающиеся статейные назначения закрываются, орган получает уведомление);
        exclude/restore — excluded с причиной (для exclude обязательна), датой и автором.
        Статусы ревью сохраняются; ministry пересчитывает триггер БД.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [action, ngr]
              properties:
                action: { type: string, enum: [transfer, exclude, restore] }
                ngr: { type: string }
                articles: { type: array, items: { type: string }, maxItems: 200, nullable: true, description: "null/нет — весь НПА" }
                target_org_id: { type: integer, description: Обязателен для transfer }
                reason: { type: string, maxLength: 500 }
      responses:
        "200": { description: "{ ok, cascaded, target } | { ok, affected }" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: JWT (7 суток), выдаётся /api/auth/login
  responses:
...
    Forbidden:
      description: Недостаточно прав / чужой орган
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Не найдено
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
...